requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.9.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.3.0.23
cached-property==1.5.2
celery==3.1.26.post2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_05nm_gqf36/croots/recipe/certifi_1663615689491/work/certifi
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-apscheduler==0.6.2
django-celery==3.3.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-mdeditor==0.1.18
django-reversion==3.0.8
django-six==1.0.4
django-stdimage==5.1.1
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
kombu==3.0.37
MarkupPy==1.14
Naked==0.1.31
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
progressbar2==4.0.0
prompt-toolkit==3.0.31
pycryptodome==3.15.0
python-utils==3.3.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
